Updated fix for TLS / delay-loading issues with Windows XP and pixman. I've taken Brent Fulgham's approach of using pthread set_specific functions instead of disabling TLS entirely, since this is safer. In summary, the problem is that __declspec(thread) compiler-directive method doesn't work with delay-loaded DLLs on pre-Vista OSes. This means XP crashes inside cairo without a fix here.
The header change here is just to force the compiler to find the pthread TLS versions before the MSVC ones.

N.B. up until now, we have been running with PIXMAN_NO_TLS defined, and not had any (known) problems with this. So we could possibly go back to using this, but it leaves us wide open to rendering problems if WebKit/cairo use pixman from multiple threads.